requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
gunicorn==21.2.0
psycopg[binary]==3.1.18
